refactor: move 'validation' to 'useValidation' hook
