+++ src/hooks/useValidation/index.ts
@@ -1,0 +1,1 @@
+export { useValidation } from './useValidation';
