--- src/utils/validation/index.ts
@@ -1,2 +1,0 @@
-export { handleValidation } from './handleValidation';
-export { validateNodeMotionProps, validateTextMotionProps } from './validation';
